+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Datasets always show geoInfo value, even if they are rasters (it will be always true)
 - Widget map creation
 - Better validations for widgetEditor
+- MapControls component added
+- BasemapControl component added
+- ShareControl component added
 
 ### Added
 - Planet pulse improvements: color used as extra dimension in more 3D layers